project restructure, added users router, get request for all users is testable and functional
