+++ api/auth/auth-service.js
@@ -1,7 +1,7 @@
 const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
 const { JWT_SECRET, JWT_EXPIRY } = require("../config");
-const User = require("../models/user-model");
+const User = require("../user/user-model");
 
 const AuthService = {
   createUser: async (newUser) => {
