--- src/Procfile
@@ -1,1 +1,0 @@
-web: node src/server.js
